+++ src/hotspot/share/runtime/os.hpp
@@ -777,7 +777,7 @@ class os: AllStatic {
   static void print_register_info(outputStream* st, const void* context);
   static bool signal_sent_by_kill(const void* siginfo);
   static void print_siginfo(outputStream* st, const void* siginfo);
-  static void print_signal_handlers(outputStream* st, char* buf, int buflen);
+  static void print_signal_handlers(outputStream* st, char* buf, size_t buflen);
   static void print_date_and_time(outputStream* st, char* buf, size_t buflen);
 
   static void print_user_info(outputStream* st);
